+++ internal/pty/session_windows.go
@@ -17,6 +17,7 @@ type winSession struct {
 	cols   int
 	rows   int
 	env    []string
+	cwd    string
 }
 
 func New() Session {
@@ -31,6 +32,10 @@ func (s *winSession) SetEnv(env []string) {
 	s.env = env
 }
 
+func (s *winSession) SetCWD(dir string) {
+	s.cwd = dir
+}
+
 func (s *winSession) Start(cmd string, args ...string) error {
 	cp, err := conpty.New(s.cols, s.rows, 0)
 	if err != nil {
@@ -39,7 +44,9 @@ func (s *winSession) Start(cmd string, args ...string) error {
 	s.cpty = cp
 
 	fullArgs := append([]string{cmd}, args...)
-	attr := &syscall.ProcAttr{}
+	attr := &syscall.ProcAttr{
+		Dir: s.cwd,
+	}
 	if len(s.env) > 0 {
 		attr.Env = append(os.Environ(), s.env...)
 	}